refactor: MVC 패턴 구조로 패키지 변경
