+++ src/main/kotlin/bridge/domain/BridgeRandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.domain
 
 import camp.nextstep.edu.missionutils.Randoms
