+++ src/main/kotlin/bridge/view/OutputView.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.view
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
